Lift up state from Activities to Trip
