+++ src/pages/Trip/Trip.jsx
@@ -1,15 +1,43 @@
 import React, { useEffect, useState } from 'react'
 import { Link } from 'react-router-dom'
-import styles from './Trip.module.css'
-import * as tripService from '../../services/tripService'
 import { useParams } from 'react-router-dom'
+
+// Styles
+import styles from './Trip.module.css'
+
+// Services
+import * as tripService from '../../services/tripService'
+import * as activityService from '../../services/activityService'
+
+// Components
 import EditTripForm from '../../components/EditTripForm/EditTripForm'
 import Activities from '../Activities/Activities'
 
 function Trip(props) {
   const { id } = useParams()
+  
   const [trip, setTrip] = useState({})
+  const [activities, setActivities] = useState([])
+  useEffect(() => {
+    const fetchTrip = async () => {
+        try {
+            const trip = await tripService.getTripById(id)
+            
+            setTrip({...trip, date: trip.date.split('T')[0]})
+            setActivities(trip.activities)
+        } catch (error) {
+            throw error
+        }
+    }
+    fetchTrip()
+    return () => { setTrip(null) }
+}, [id])
+  
   const [showEditTripForm, setShowEditTripForm] = useState(false)
+  
+  const handleToggle = () => {
+    setShowEditTripForm(!showEditTripForm)
+  }
 
   const handleUpdateTrip = async (updatedTripData) => {
     try {
@@ -27,23 +55,18 @@ function Trip(props) {
     } 
   }
 
-  useEffect(() => {
-    const fetchTrip = async () => {
-        try {
-            const trip = await tripService.getTripById(id)
-            
-            setTrip({...trip, date: trip.date.split('T')[0]})
-        } catch (error) {
-            throw error
-        }
-    }
-    fetchTrip()
-    return () => { setTrip(null) }
-}, [id])
+  const handleAddActivity = async (newActivityData) => {
+    const newActivity = await activityService.create(newActivityData);
+    await tripService.update({$push: {activities: newActivity._id}}, id)
+    setActivities([newActivity, ...activities]);
+  }
 
-    const handleToggle = () => {
-      setShowEditTripForm(!showEditTripForm)
-    }
+  const handleDeleteActivity = id => {
+    activityService.deleteOne(id)
+      .then(
+        setActivities(activities.filter(activity => id !== activity._id))
+      ) 
+  }
 
   return (
     <div className={styles.container}>
@@ -51,7 +74,9 @@ function Trip(props) {
       <h2>{trip.notes}</h2>
       <h3>{trip.date}</h3>
       <Activities
-        tripId={id}
+        activities={activities}
+        handleAddActivity={handleAddActivity}
+        handleDeleteActivity={handleDeleteActivity}
       />
       {/* <Link to={'/activities'}>Activities</Link>
       <Link to={'/checklist'}>Checklists</Link> */}
